modifies configuration to send mail via SMTP
Previously this service was using Amazon's SES service to send emails.
It was deemed that SendGrid should be used instead. This makes config
changes to direct the SMTP request to SendGrid and populates the
username field with apiKey in Base64 as required by SendGrid.

The ApiKey in Base64 for SendGrid is also expected to be present
in the prod environment